feat: add main service

+++ main-service/src/main/java/ru/practicum/ewm/category/CategoryServiceImpl.java
@@ -2,37 +2,73 @@ package ru.practicum.ewm.category;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.ewm.event.Event;
+import ru.practicum.ewm.event.EventRepository;
+import ru.practicum.ewm.exception.ConflictOperationException;
 import ru.practicum.ewm.exception.EntityNotFoundException;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Transactional
     @Override
-    public Category add(Category category) {
-        return categoryRepository.save(category);
+    public Category add(Category newCategory) {
+        List<Category> categoriesByName = categoryRepository.findAll().stream()
+                .filter(x -> x.getName().toLowerCase().equals(newCategory.getName().toLowerCase()))
+                .collect(Collectors.toList());
+        if (!categoriesByName.isEmpty()) {
+            throw new ConflictOperationException("Категория с наименованием = " + newCategory.getName() + " уже существует");
+        }
+        return categoryRepository.save(newCategory);
     }
 
     @Transactional
     @Override
-    public Category update(Category category) {
-        categoryRepository.findById(category.getId()).orElseThrow(() -> {
-            throw new EntityNotFoundException("Категория с id = " + category.getId() + " не существует");
-        });
-        return categoryRepository.save(category);
+    public Category update(Category newCategory) {
+        List<Category> categoriesByName = categoryRepository.findAll().stream()
+                .filter(x -> x.getName().toLowerCase().equals(newCategory.getName().toLowerCase())
+                        && !x.getId().equals(newCategory.getId()))
+                .collect(Collectors.toList());
+        if (!categoriesByName.isEmpty()) {
+            throw new ConflictOperationException("Категория с наименованием = " + newCategory.getName() + " уже существует");
+        }
+        return categoryRepository.save(newCategory);
     }
 
     @Transactional
     @Override
     public void delete(Long id) {
-        categoryRepository.findById(id).orElseThrow(() -> {
-            throw new EntityNotFoundException("Категория с id = " + id + " не существует");
-        });
+        List<Event> eventsByCategory = eventRepository.findByCategory(getById(id));
+        if (!eventsByCategory.isEmpty()) {
+            throw new ConflictOperationException("Удаление невозможно, так как существуют события по данной категории");
+        }
         categoryRepository.deleteById(id);
+    }
+
+    @Override
+    public List<Category> getAll(Integer from, Integer size) {
+        Pageable page = PageRequest.of(from / size, size, Sort.by("id").ascending());
+        Page<Category> all = categoryRepository.findAll(page);
+        return all.getContent();
+    }
+
+    @Override
+    public Category getById(long catId) {
+        return categoryRepository.findById(catId).orElseThrow(() -> {
+            throw new EntityNotFoundException("Категория с id = " + catId + " не существует");
+        });
     }
 }
